rewrite

+++ internal/database/migrations/20240522150321_device.sql
@@ -2,7 +2,7 @@
 -- +goose StatementBegin
 CREATE TABLE devices (
     imei VARCHAR(15) PRIMARY KEY,
-    battery_power SMALLINT NOT NULL,
+    battery_power INT NOT NULL,
     charging BOOLEAN NOT NULL,
     last_status_packet TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
