+++ internal/database/migrations/20240522150349_positions.sql
@@ -2,10 +2,10 @@
 -- +goose StatementBegin
 CREATE TABLE positions (
     id SERIAL PRIMARY KEY,
-    latitude float8 NOT NULL,
-    longitude float8 NOT NULL,
-    speed SMALLINT NOT NULL,
-    heading SMALLINT NOT NULL,
+    latitude REAL NOT NULL,
+    longitude REAL NOT NULL,
+    speed INT NOT NULL,
+    heading INT NOT NULL,
     device_imei VARCHAR(15) NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (device_imei) REFERENCES devices(imei)
